Prep work before adding each db recipe
Brought in openstack-common as a dep.  Updated the tests to
better match what the recipes are doing (added tests where
missing).  Updated attributes to match the format of our other
cookbooks.  This is mostly some cleanups, to prepare for features.
Since this cookbook is not in use, made a larger commit than
would have if adding functionality to a released cookbook.  Also,
upgraded to new Berkshelf which handles locking to cookbooks properly.

Change-Id: I8e4123d30012a9890ce3029184e1396d2083a6f1