Rework deletion of insecure users
Due to a bug in MySQL [1] -- or possibly the Ubuntu MySQL server
packages -- issuing a call to DROP USER fails consistently when
provisioning a node with the recipes/mysql-server.rb recipe.

This reworks the DROP USER to a DELETE call on the mysql.user
table. Note that the DROP USER is preferred over the call to
manually DELETE FROM mysql.user because Galera cluster does not
replicate DELETE events on MyISAM tables but does replicate
DROP USER events [2]. So, when the MySQL bug is fixed or a
workaround is found, we should revert this patch.

[1] http://bugs.mysql.com/bug.php?id=69644
[2] http://www.codership.com/wiki/doku.php?id=limitations

Change-Id: I0fab1a0f85d55eab96984064131d14a21b2236e3