Fix issue with mysql not restarting on new bind_ip - RHEL/CENTOS
Addresses: bug #1277041
Implements: blueprint centos-rhel-for-ops-database

By default mysql binds to 127.0.0.1. Setting the bind_ip attribute to
something else, as we do in the ops-database cookbook before calling the
mysql::server recipe, causes the my.cnf to be populated with the correct IP.

However, because of this bug in the upstream mysql cookbook:
https://tickets.opscode.com/browse/COOK-4161
mysql does not get restarted (it only gets reloaded)
and so the new bind_ip does not take effect.

This commit adds a workaround until the upstream bug is fixed. The
workaround directly modifies the template resource from the resource
collection to add a 'restart mysql' notification.

** This particular patchset builds on the last by adding support for
rhel/centos which use a different name for the my.cnf template resource **

Change-Id: I357bf1458a1781189922032cdc8da8ad9e0a03a7