Increase default connect timeout for mariadb
The mariadb cookbook imposes a default connect_timeout of 5 seconds.
This is too aggressive for services to initialize, and they error as a
result. To avoid such problems, mariadb should allow for a higher
threshold before timing out a connecting service. This change raises
that timeout to 30 seconds to be more graceful for services warming up.

Change-Id: I70775929dc49ed8c00a23bc7e354ebf9e9feb7f0